requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
DateTime==5.5
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
openpyxl==3.1.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
yarl==1.9.4
zope.interface==6.2

phonenumbers~=8.13.34
aioschedule~=0.5.2